+++ src/gppc/__description__.py
@@ -1,3 +1,5 @@
-"""Contains description info."""
-
+"""DESCRIPTION"""
+__title__ = "gppc"
+__author__ = "moxxos"
+__copyright__ = "Copyright (C) 2022 moxxos"
 short_description = "Check OSRS Grand Exchange prices from the command line. Includes limited module functionality."